+++ webui/react/src/ioTypes.ts
@@ -8,7 +8,6 @@ import {
   HyperparameterType,
   LogLevel,
   RunState,
-  ValueOf,
 } from 'types';
 import { DetError, ErrorLevel, ErrorType } from 'utils/error';
 
@@ -34,19 +33,6 @@ export const optional = (x: io.Mixed): io.Mixed | io.NullC | io.UndefinedC => {
 export const nullable = (x: io.Mixed): io.Mixed | io.NullC => {
   return io.union([x, io.null]);
 };
-
-export class ValueofType<D extends { [key: string]: unknown }> extends io.Type<ValueOf<D>> {
-  readonly _tag: 'ValueofType' = 'ValueofType' as const;
-  constructor(
-    name: string,
-    is: ValueofType<D>['is'],
-    validate: ValueofType<D>['validate'],
-    encode: ValueofType<D>['encode'],
-    readonly values: D,
-  ) {
-    super(name, is, validate, encode);
-  }
-}
 
 class Float extends io.Type<number, number | string, unknown> {
   readonly _tag: 'FloatType' = 'FloatType' as const;
@@ -84,26 +70,6 @@ class Float extends io.Type<number, number | string, unknown> {
 }
 
 export const float = new Float();
-
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface ValueofC<D extends { [key: string]: unknown }> extends ValueofType<D> {}
-
-export function valueof<D extends { [key: string]: unknown }>(
-  values: D,
-  name: string = Object.values(values)
-    .map((k) => JSON.stringify(k))
-    .join(' | '),
-): ValueofC<D> {
-  const valueSet = new Set(Object.values(values));
-  const is = (u: unknown): u is ValueOf<D> => valueSet.has(u);
-  return new ValueofType(
-    name,
-    is,
-    (u, c) => (is(u) ? io.success(u) : io.failure(u, c)),
-    io.identity,
-    values,
-  );
-}
 
 /* Slot */
